Move python/tlm_adjoint_firedrake/ to tlm_adjoint/firedrake/
