+++ docs/manual.tex
@@ -373,7 +373,7 @@ from tlm_adjoint.fenics import *
 The Firedrake backend may be used via, e.g.
 \begin{lstlisting}
 from firedrake import *
-from tlm_adjoint_firedrake import *
+from tlm_adjoint.firedrake import *
 \end{lstlisting}
 
 The NumPy backend may be used via, e.g.
