+++ examples/firedrake/diffusion/diffusion.py
@@ -19,8 +19,8 @@
 # along with tlm_adjoint.  If not, see <https://www.gnu.org/licenses/>.
 
 from firedrake import *
-from tlm_adjoint_firedrake import *
-from tlm_adjoint_firedrake import manager as _manager
+from tlm_adjoint.firedrake import *
+from tlm_adjoint.firedrake import manager as _manager
 
 # import h5py
 import mpi4py.MPI as MPI
